Added first go at autogeneration of ei and rebin
autoEi gets a rough value of ei from monitor spectrum, using peakdet.py
a simple peak find routine. This seems more robust than FindPeaks
algorithum which fails to find real peaks in a reproducable manner.
mari_auto_example.py is an example of useage for a couple of runs.
The ei gues is hard coded to look at spectrum 2 which in any right
minded instrument read het mari and maps is the monitor after the
fermi, this is not the case for merlin and let where the monitors are
located at the end of the mapping file.